+++ packages/http/src/__tests__/http-prism-instance.spec.ts
@@ -1,6 +1,7 @@
 import { createLogger, IPrism } from '@stoplight/prism-core';
 import { DiagnosticSeverity } from '@stoplight/types';
 import { IHttpOperation } from '@stoplight/types';
+import { Scope as NockScope } from 'nock';
 import * as nock from 'nock';
 import { basename, resolve } from 'path';
 import { createInstance, IHttpConfig, IHttpRequest, IHttpResponse, ProblemJsonError } from '../';
@@ -16,8 +17,34 @@ const staticExamplesOas2Path = fixturePath('static-examples.oas2.json');
 const serverValidationOas2Path = fixturePath('server-validation.oas2.json');
 const serverValidationOas3Path = fixturePath('server-validation.oas3.json');
 
+const { version: prismVersion } = require('../../package.json');
+
+type Prism = IPrism<IHttpOperation, IHttpRequest, IHttpResponse, IHttpConfig, { path: string }>;
+type NockResWithInterceptors = NockScope & { interceptors: Array<{ req: { headers: string[] } }> };
+
+async function checkUserAgent(config: IHttpConfig, prism: Prism, headers = {}) {
+  const oasBaseUrl = 'http://example.com/api';
+
+  const nockResult = nock(oasBaseUrl)
+    .get('/pet')
+    .reply(200);
+
+  await prism.process(
+    {
+      method: 'get',
+      url: {
+        path: '/pet',
+      },
+      headers,
+    },
+    config,
+  );
+
+  return (nockResult as NockResWithInterceptors).interceptors['0'].req.headers['user-agent'];
+}
+
 describe('Http Client .process', () => {
-  let prism: IPrism<IHttpOperation, IHttpRequest, IHttpResponse, IHttpConfig, { path: string }>;
+  let prism: Prism;
 
   describe.each`
     specName                              | specPath
@@ -170,6 +197,26 @@ describe('Http Client .process', () => {
           expect(result.output).toBeDefined();
           expect(result.output!.statusCode).toEqual(200);
           expect(result.output!.body).toEqual(reply);
+        });
+      });
+
+      describe('Prism user-agent header', () => {
+        describe('when the defaults are used', () => {
+          it('should use Prism/<<version>> for the header', async () => {
+            const userAgent = await checkUserAgent(config, prism);
+
+            expect(userAgent).toBe(`Prism/${prismVersion}`);
+          });
+        });
+
+        describe('when user-agent is being overwritten', () => {
+          it('should have user specified string as the header', async () => {
+            const userAgent = await checkUserAgent(config, prism, {
+              'user-agent': 'Other_Agent/1.0.0',
+            });
+
+            expect(userAgent).toBe('Other_Agent/1.0.0');
+          });
         });
       });
     });